+++ README.md
@@ -21,7 +21,7 @@
 - directions to install
 
 ```bash
-pip install SeleniumCookie==0.1
+pip install SeleniumCookie==0.2
 ```
 
 - directions to execute
@@ -62,11 +62,11 @@ driver.get(url)
 
 - Checking for added Cookies
 
-	- Run the above script
-	- Now in Chrome Search Bar: chrome://settings/siteData
-	- Added Cookies can be seen now
-	- Additon of cookie subject to it's various parameters and data it has
-	- In my test 279 (119 websites) cookies were injected and 2924 were rejected
+      	- Run the above script
+      	- Now in Chrome Search Bar: chrome://settings/siteData
+      	- Added Cookies can be seen now
+      	- Additon of cookie subject to it's various parameters and data it has
+      	- In my test 279 (119 websites) cookies were injected and 2924 were rejected
 
 </br>
