+++ SeleniumCookie/tests/__init__.py
@@ -1,0 +1,4 @@
+try:
+	from tests import tests
+except:
+	import tests
